Added cast from generic to primitive for java. (Please run svn up && mvn clean test before committing)

+++ sigmah/src/main/java/org/sigmah/shared/dto/report/ReportModelDTO.java
@@ -47,7 +47,7 @@ public class ReportModelDTO extends BaseModelData implements EntityDTO {
     
     public int getOrganizationId()
     {
-    	return get("organizationId");
+    	return (Integer)get("organizationId");
     }
     
     public void setOrganizationId(int organizationId)
